Nailed httplib2 to a version with working SSL CRL checking.
CRL checking went completely broken when httplib2 went to 0.7.0; it's not clear how to work around that with the current version, so I'm pinning it to an older version to recover the behaviour we had before. I'm aware that this might introduce security issues; I'll get a security-oriented dude to help out with a proper fix at some point.

See https://github.com/freshbooks/refreshbooks/issues/6 for further information.